DPR2-239: Rename domain.model to datahub.model as per code review request.
