+++ src/it/java/uk/gov/justice/digital/client/s3/S3DataProviderIT.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.io.TempDir;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.justice.digital.config.JobArguments;
-import uk.gov.justice.digital.domain.model.SourceReference;
+import uk.gov.justice.digital.datahub.model.SourceReference;
 import uk.gov.justice.digital.test.BaseMinimalDataIntegrationTest;
 
 import java.io.IOException;
