+++ src/main/java/uk/gov/justice/digital/service/SourceReferenceService.java
@@ -12,7 +12,7 @@ import lombok.val;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import uk.gov.justice.digital.client.s3.S3SchemaClient;
 import uk.gov.justice.digital.converter.avro.AvroToSparkSchemaConverter;
-import uk.gov.justice.digital.domain.model.SourceReference;
+import uk.gov.justice.digital.datahub.model.SourceReference;
 
 import java.util.*;
 import java.util.stream.Collectors;
