+++ src/main/java/uk/gov/justice/digital/zone/Zone.java
@@ -3,7 +3,7 @@ package uk.gov.justice.digital.zone;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SparkSession;
-import uk.gov.justice.digital.domain.model.SourceReference;
+import uk.gov.justice.digital.datahub.model.SourceReference;
 import uk.gov.justice.digital.exception.DataStorageException;
 
 public interface Zone {
