+++ src/test/java/uk/gov/justice/digital/datahub/model/TableIdentifierTest.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.domain.model;
+package uk.gov.justice.digital.datahub.model;
 
 import lombok.val;
 import org.junit.jupiter.api.Test;
