+++ src/test/java/uk/gov/justice/digital/service/DataStorageServiceTest.java
@@ -22,8 +22,8 @@ import org.mockito.MockedStatic;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.justice.digital.config.BaseSparkTest;
 import uk.gov.justice.digital.config.JobArguments;
-import uk.gov.justice.digital.domain.model.SourceReference;
-import uk.gov.justice.digital.domain.model.TableIdentifier;
+import uk.gov.justice.digital.datahub.model.SourceReference;
+import uk.gov.justice.digital.datahub.model.TableIdentifier;
 import uk.gov.justice.digital.exception.DataStorageException;
 import uk.gov.justice.digital.exception.DataStorageRetriesExhaustedException;
