+++ src/test/java/uk/gov/justice/digital/test/MinimalTestData.java
@@ -11,7 +11,7 @@ import org.apache.spark.sql.types.Metadata;
 import org.apache.spark.sql.types.StructField;
 import org.apache.spark.sql.types.StructType;
 import uk.gov.justice.digital.common.CommonDataFields;
-import uk.gov.justice.digital.domain.model.SourceReference;
+import uk.gov.justice.digital.datahub.model.SourceReference;
 
 import java.util.Arrays;
 import java.util.List;
